Support GPUs request in Kubernetes cluster
The admin can tell us the GPU vendor (Nvidia or AMD) when they are configuring a Kubernetes server.
When a container needs a GPU, the command author will include "generic-resources": {"gpu": "n"}, where n is a string-encoded integer number of GPUs.

If the container to be launched requests GPUs, we can turn the vendor into a string (ex. nvidia.com/gpu: 1 ) that the Kubernetes API expects. If this GPU vendor info is null, we are going to throw an exception.